pyproject.toml
[project]
name = "food-online-receipt"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "langchain>=0.3.27",
    "langchain-google-genai>=2.1.10",
    "pillow>=11.3.0",
    "pydantic>=2.11.7",
    "pytesseract>=0.3.13",
    "sqlalchemy>=2.0.43",
    "streamlit>=1.49.1",
]
